openssl, io: Fix linux compatibility

+++ src/coap_session.c
@@ -238,7 +238,7 @@ void coap_session_send_csm(coap_session_t *session) {
     coap_session_disconnected(session, COAP_NACK_NOT_DELIVERABLE);
   } else {
     ssize_t bytes_written = coap_session_send_pdu(session, pdu);
-    if (bytes_written != pdu->used_size + pdu->hdr_size)
+    if (bytes_written != (ssize_t)pdu->used_size + pdu->hdr_size)
       coap_session_disconnected(session, COAP_NACK_NOT_DELIVERABLE);
   }
   if (pdu)
